Refactor device detection logic
Another step in the major refactoring of the device management.
Added device manager class.
I tried to minimize this commit but it is still rather big. I kept the names of files such as detect.py, although the name and functionality has changed. There's still a lot or room to refactor the code. This is merely the first step.