Fix CLI printing
In general, DymoLabeler and rendering classes are tightly coupled with the selected device (which incluides current physical tape installed in it).
We need to select the device first, so that preview or printing will match the intended device.
In the future we might want to preview labels for device that is offline/absent, but for now will do with requiring a connected working device.