Fix issues with the display of permissions instructions
One subtle issue is that when printing the exception, line breaks were
being inserted according to the terminal width. To avoid this, I broke the
command into short lines.

The error now reads:

```
UsbDeviceError:

You do not have sufficient access to the device. You probably want to add the a udev
rule in /etc/udev/rules.d with the following command:

    echo 'ACTION=="add",'\
         'SUBSYSTEMS=="usb",'\
         'ATTRS{idVendor}=="0922",'\
         'ATTRS{idProduct}=="1001",'\
         'MODE="0666"' \
      | sudo tee /etc/udev/rules.d/91-labelle-0922-1001.rules

Next, refresh udev with:

    sudo udevadm control --reload-rules
    sudo udevadm trigger --attr-match=idVendor="0922"

Finally, turn your device off and back on again to activate the new permissions.

If this still does not resolve the problem, you might need to reboot.
In case rebooting is necessary, please report this at <https://github.com/labelle-org/labelle/issues/6>.
We are still trying to figure out a simple procedure which works for everyone.
In case you still cannot connect, or if you have any information or ideas, please
post them at that link.
```