Add new --batch option to read batch commands from stdin
Makes it possible to generate more complicated multi-block layouts in
the CLI, akin to the GUI.  I started this because I wanted the QR code
after, rather than before, the text.

For now, I define only text and qr commands.  If the maintainers accept
this patch, I will round it out with more commands.